Fix INT16_MIN missing + -Wsign-compare + -Wreorder
Former-commit-id: 4f72c6e7be5d37dbdd855ef046e05e570d91aea2
Former-commit-id: 0ad91ce462c90fe7a872112d73516493a3e596fb [formerly 4c2ce29d3bb0403480f9e57240e61fe5e272a4ee]
Former-commit-id: caca51581e5f70eeb2b5a8b833a6cd9855bb38f6
Former-commit-id: 7640b5032e8353e703c464787b602db6c93f08eb [formerly 4bbed385a10705c7ae8c36cf04f8bc75ee0c32df]
Former-commit-id: 107bbab179167e11ddc2bb04db49bcf29efd12d7
Former-commit-id: 7bd0092050f62813d7df40e4922cece8b97cf93f
Former-commit-id: afbf4845619e7831612eba7d2e6886b4fb90717e

+++ src/TreeSegmentManager.cpp
@@ -160,7 +160,7 @@ TreeSegment TreeSegmentManager::build_combination(std::vector<TreeSegment> &tree
 {
   TreeSegment tree = trees[combination[0]];
 
-  for (int i = 1 ; i < combination.size() ; i++)
+  for (size_t i = 1 ; i < combination.size() ; i++)
   {
     int id = combination[i];
     TreeSegment& tree2 = trees[id];
@@ -174,7 +174,7 @@ std::vector<TreeSegment> TreeSegmentManager::get_non_combined_tree(std::vector<T
 {
   std::vector<TreeSegment> out_trees;
 
-  for (int i = 0 ; i < trees.size() ; i++)
+  for (size_t i = 0 ; i < trees.size() ; i++)
   {
     if(std::find(combination.begin(), combination.end(), i) == combination.end())
       out_trees.push_back(trees[i]);
@@ -251,7 +251,7 @@ void TreeSegmentManager::planimetric_filter(std::vector<PointXYZ> &subProfile, s
   double threshold = meanValueForThreshold + 2*stdValueForThreshold;
   //--------------------------------------------------------------------------------------
   // Keeping all points below threshold and storage of their IDs
-  int i = 0, keep = 0;
+  size_t i = 0, keep = 0;
   subProfileSubset.push_back(subProfile[0]);
   while ( (i < subProfile.size()-1) && (dist[i] <= threshold) )
   {
@@ -262,7 +262,7 @@ void TreeSegmentManager::planimetric_filter(std::vector<PointXYZ> &subProfile, s
 
 Rcpp::List TreeSegmentManager::to_R()
 {
-  int n = treeStorage.size();
+  size_t n = treeStorage.size();
 
   std::replace(points_labels.begin(), points_labels.end(), 0, NA_INTEGER);
